exp/go-templates: improved WASM templating logic
Various teaks to make things more configurable and dynamic, with hashing
and resource-based management of js and WASM files to the extent
possible.
