+++ layouts/partials/footer.html
@@ -31,7 +31,13 @@
 
 <!-- Go WASM -->
 {{ if .Params.wasmEnabled }}
-<script src="/exp/wasm_exec.js"></script>
+<!-- Import and hash -->
+{{ $wasmExec := resources.Get "/js/exp/wasm/wasm_exec.js" |
+resources.Fingerprint "sha256" }}
+<script
+  src="{{ $wasmExec.RelPermalink }}"
+  integrity="{{ $wasmExec.Data.Integrity }}"
+></script>
 {{ end }}
 
 <script>
