+++ layouts/partials/jsx.html
@@ -1,3 +1,6 @@
+<!-- NOTE: Prettier is disabled for this file to allow for better formatting of
+  the template logic.. -->
+
 <!-- JS Build defines -->
 {{ $defines := dict "process.env.NODE_ENV" hugo.Environment }}
 <!-- JS Build options. the JSX* options are key to using the lighter weight
@@ -5,16 +8,25 @@
 specifically. -->
 {{ $options := dict "defines" $defines "JSXFactory" "h" "JSXFragment" "Fragment"
 "minify" true }}
-<!-- per-environment customizations -->
-{{ if eq hugo.Environment "development" }}
+
 <!-- In non-production, enable source maps -->
 <!-- TODO: Enable always when options other than "inline" are available -->
-{{ $options = merge $options (dict "sourcemap" "inline") }} {{ end }}
-<!-- Search logic -->
-{{ $js := resources.Get . | js.Build $options }}
+{{ if eq hugo.Environment "development" }}
+  {{ $options = merge $options (dict "sourcemap" "inline") }}
+{{ end }}
+
+<!-- Get raw JS file -->
+{{ $js := resources.Get . }}
+<!-- Optionally, render template data within the file -->
+{{ if strings.Count ".tpl" . }}
+  {{ $jsName := strings.Replace . ".tpl" "" }}
+  {{ $js = resources.Get . | resources.ExecuteAsTemplate $jsName . }}
+{{ end }}
+<!-- Build JS file -->
+{{ $js = $js | js.Build $options }}
 <!-- SRI hashing: https://gohugo.io/hugo-pipes/fingerprint/ -->
-{{ $secureJS := $js | resources.Fingerprint "sha256" }}
+{{ $js = $js | resources.Fingerprint "sha256" }}
 <script
-  src="{{ $secureJS.RelPermalink }}"
-  integrity="{{ $secureJS.Data.Integrity }}"
+  src="{{ $js.RelPermalink }}"
+  integrity="{{ $js.Data.Integrity }}"
 ></script>
